+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 #### Fixed
 - Fixed the width of showed tagged idps in case the count of idps is equal to (x * 3) + 1
 - Using try{}catch{} to avoid to PerunException in PerunEntitlement.php
+- Return [] instead of null in getFacilityCapability via RPC, if facilityCapability is not set
 
 ## [v3.8.0]
 #### Changed